Single entry point into training
